Refactor algorithms into individual modules
