+++ src/algorithms/bucket-sort.js
@@ -1,0 +1,19 @@
+function bucketSort(arr, bucketSize = 5) {
+  if (arr.length === 0) return [];
+  const min = Math.min(...arr);
+  const max = Math.max(...arr);
+  const bucketCount = Math.floor((max - min) / bucketSize) + 1;
+  const buckets = Array.from({ length: bucketCount }, () => []);
+  for (const num of arr) {
+    const idx = Math.floor((num - min) / bucketSize);
+    buckets[idx].push(num);
+  }
+  const result = [];
+  for (const bucket of buckets) {
+    bucket.sort((a, b) => a - b);
+    result.push(...bucket);
+  }
+  return result;
+}
+
+module.exports = bucketSort;
