+++ src/algorithms/find-peak-element.js
@@ -1,0 +1,15 @@
+function findPeakElement(arr) {
+  let l = 0,
+    r = arr.length - 1;
+  while (l < r) {
+    const m = Math.floor((l + r) / 2);
+    if (arr[m] < arr[m + 1]) {
+      l = m + 1;
+    } else {
+      r = m;
+    }
+  }
+  return l;
+}
+
+module.exports = findPeakElement;
