+++ src/algorithms/frequency-map-counting.js
@@ -1,0 +1,9 @@
+function frequencyMapCounting(str) {
+  const map = {};
+  for (const ch of str) {
+    map[ch] = (map[ch] || 0) + 1;
+  }
+  return map;
+}
+
+module.exports = frequencyMapCounting;
