+++ src/algorithms/kmp-prefix-function.js
@@ -1,0 +1,14 @@
+function kmpPrefixFunction(s) {
+  const pi = new Array(s.length).fill(0);
+  for (let i = 1; i < s.length; i++) {
+    let j = pi[i - 1];
+    while (j > 0 && s[i] !== s[j]) {
+      j = pi[j - 1];
+    }
+    if (s[i] === s[j]) j++;
+    pi[i] = j;
+  }
+  return pi;
+}
+
+module.exports = kmpPrefixFunction;
