+++ src/algorithms/rotated-array-search.js
@@ -1,0 +1,23 @@
+function rotatedArraySearch(arr, target) {
+  let l = 0, r = arr.length - 1;
+  while (l <= r) {
+    const m = Math.floor((l + r) / 2);
+    if (arr[m] === target) return m;
+    if (arr[l] <= arr[m]) {
+      if (target >= arr[l] && target < arr[m]) {
+        r = m - 1;
+      } else {
+        l = m + 1;
+      }
+    } else {
+      if (target > arr[m] && target <= arr[r]) {
+        l = m + 1;
+      } else {
+        r = m - 1;
+      }
+    }
+  }
+  return -1;
+}
+
+module.exports = rotatedArraySearch;
