+++ src/algorithms/run-length-encoding.js
@@ -1,0 +1,16 @@
+function runLengthEncode(str) {
+  if (str.length === 0) return '';
+  let out = '';
+  let count = 1;
+  for (let i = 1; i <= str.length; i++) {
+    if (i < str.length && str[i] === str[i - 1]) {
+      count++;
+    } else {
+      out += str[i - 1] + count;
+      count = 1;
+    }
+  }
+  return out;
+}
+
+module.exports = runLengthEncode;
